Add abstract ConeRayEstimator base class and refactor estimators
Introduce an abstract base class ConeRayEstimator to unify common
logic for cone ray angle estimation components. Refactor
GuidedConeRayEstimator and OnGestureConeRayEstimator to inherit from
this base.
