+++ Runtime/Components/ConeRayAnglesEstimator/GuidedConeRayEstimator.cs
@@ -1,10 +1,5 @@
 using System.Collections.Generic;
-using UnityEngine;
-using System;
-using UnityEngine.Assertions;
 using System.Linq;
-using System.Threading.Tasks;
-using System.Collections;
 using ubco.ovilab.HPUI.Interaction;
 
 namespace ubco.ovilab.HPUI.Components
@@ -17,52 +12,14 @@ namespace ubco.ovilab.HPUI.Components
     /// of interaction per ray. Makes use of <see cref="HPUIInteractorConeRayAngleSegment"/>
     /// from <see cref="OnGestureConeRayEstimator"/> for the list of phalanges.
     /// </summary>
-    public class GuidedConeRayEstimator
+    public class GuidedConeRayEstimator: ConeRayEstimator
     {
         private bool isCalibrationActive = false;
 
         public bool IsCalibrationActive { get => isCalibrationActive; set => isCalibrationActive = value; }
 
-        private HPUIInteractor interactor;
-        private HPUIFullRangeRayCastDetectionLogic fullRayDetectionLogic;
-        private HPUIInteractorFullRangeAngles fullRangeAngles;
-        private List<InteractionDataRecord> interactionRecords = new();
-        private List<List<HPUIRayCastDetectionBaseLogic.RaycastDataRecord>> currentInteractionData = new();
-
-        public GuidedConeRayEstimator(HPUIInteractor interactor)
-        {
-
-            if (!(interactor.DetectionLogic is HPUIFullRangeRayCastDetectionLogic fullRayDetectionLogic))
-            {
-                throw new ArgumentException("Interactor is expected to have `HPUIFullRangeRayCastDetectionLogic` as the DetectionLogic.");
-            }
-            this.interactor = interactor;
-            this.fullRayDetectionLogic = fullRayDetectionLogic;
-            this.fullRangeAngles = fullRayDetectionLogic.FullRangeRayAngles;
-
-            fullRayDetectionLogic.raycastData += RaycastDataCallback;
-        }
-
-        /// <summary>
-        /// The callback used to get the data from the <see cref="HPUIFullRangeRayCastDetectionLogic.raycastData"/>.
-        /// </summary>
-        protected void RaycastDataCallback(List<HPUIRayCastDetectionBaseLogic.RaycastDataRecord> raycastDataRecords)
-        {
-            // ensuring the provided interactor is the same as the one providing the callbacks.
-            if (!isCalibrationActive)
-            {
-                return;
-            }
-
-            Assert.AreEqual(fullRangeAngles,
-                            ((HPUIFullRangeRayCastDetectionLogic)interactor.DetectionLogic).FullRangeRayAngles,
-                            $"Interactor {fullRangeAngles.name} is not the same as {((HPUIFullRangeRayCastDetectionLogic)interactor.DetectionLogic).FullRangeRayAngles.name}");
-
-            if (raycastDataRecords.Count > 0)
-            {
-                currentInteractionData.Add(raycastDataRecords);
-            }
-        }
+        public GuidedConeRayEstimator(HPUIInteractor interactor) : base(interactor)
+        { }
 
         /// <summary>
         /// The callback used with the interactable gesture event to track the events.
@@ -74,125 +31,8 @@ namespace ubco.ovilab.HPUI.Components
             currentInteractionData = new();
         }
 
-        /// <summary>
-        /// Stops the data collection and initates the estimation. Once done, callback will be invoked with the estimated asset.
-        /// </summary>
-        /// <remarks>
-        /// This will unsubscribe to <see cref="HPUIInteractor.DetectionLogic"/>interactor.DetectionLogic</see>
-        /// and the <see cref="IHPUIInteractable.GestureEvent">GestureEvent</see> of each
-        /// interactable it is tracking.
-        /// </remarks>
-        public void EstimateConeRayAngles(Action<HPUIInteractorConeRayAngles> callback)
-        {
-            HPUIInteractorConeRayAngles estimatedConeRayAngles = ScriptableObject.CreateInstance<HPUIInteractorConeRayAngles>();
-
-            fullRayDetectionLogic.raycastData -= RaycastDataCallback;
-
-            interactor.StartCoroutine(EstimationCoroutine(callback, estimatedConeRayAngles));
-        }
-
-        /// <summary>
-        /// Coroutine that does the actual work of <see cref="EstimateConeRayAngles"/>.
-        /// </summary>
-        protected IEnumerator EstimationCoroutine(Action<HPUIInteractorConeRayAngles> callback, HPUIInteractorConeRayAngles estimatedConeRayAngles)
-        {
-            yield return null;
-            HPUIInteractorConeRayAngleSegment[] segments = (HPUIInteractorConeRayAngleSegment[])Enum.GetValues(typeof(HPUIInteractorConeRayAngleSegment));
-
-            Task<List<HPUIInteractorRayAngle>>[] tasks = new Task<List<HPUIInteractorRayAngle>>[segments.Length];
-
-            int i = 0;
-            foreach (HPUIInteractorConeRayAngleSegment segment in segments)
-            {
-                tasks[i++] = Task.Run(() => EstimateConeAnglesForSegment(segment, interactionRecords));
-            }
-
-            while (tasks.Any(t => !t.IsCompleted))
-            {
-                IEnumerable<Exception> exceptions = tasks.Select(t => t.Exception).Where(t => t != null);
-                if (exceptions.Count() != 0)
-                {
-                    // Throwing the first thing that comes through.
-                    // MAYBE: Should all of them be processed somehow?
-                    throw exceptions.First();
-                }
-                yield return null;
-            }
-
-            for (i = 0; i < segments.Length; ++i)
-            {
-                HPUIInteractorConeRayAngleSegment segment = segments[i];
-                List<HPUIInteractorRayAngle> coneAnglesForSegment = tasks[i].Result;
-                switch (segment)
-                {
-                    case HPUIInteractorConeRayAngleSegment.IndexDistalSegment:
-                        estimatedConeRayAngles.IndexDistalAngles = coneAnglesForSegment;
-                        break;
-                    case HPUIInteractorConeRayAngleSegment.IndexIntermediateSegment:
-                        estimatedConeRayAngles.IndexIntermediateAngles = coneAnglesForSegment;
-                        break;
-                    case HPUIInteractorConeRayAngleSegment.IndexProximalSegment:
-                        estimatedConeRayAngles.IndexProximalAngles = coneAnglesForSegment;
-                        break;
-
-                    case HPUIInteractorConeRayAngleSegment.MiddleDistalSegment:
-                        estimatedConeRayAngles.MiddleDistalAngles = coneAnglesForSegment;
-                        break;
-                    case HPUIInteractorConeRayAngleSegment.MiddleIntermediateSegment:
-                        estimatedConeRayAngles.MiddleIntermediateAngles = coneAnglesForSegment;
-                        break;
-                    case HPUIInteractorConeRayAngleSegment.MiddleProximalSegment:
-                        estimatedConeRayAngles.MiddleProximalAngles = coneAnglesForSegment;
-                        break;
-
-                    case HPUIInteractorConeRayAngleSegment.RingDistalSegment:
-                        estimatedConeRayAngles.RingDistalAngles = coneAnglesForSegment;
-                        break;
-                    case HPUIInteractorConeRayAngleSegment.RingIntermediateSegment:
-                        estimatedConeRayAngles.RingIntermediateAngles = coneAnglesForSegment;
-                        break;
-                    case HPUIInteractorConeRayAngleSegment.RingProximalSegment:
-                        estimatedConeRayAngles.RingProximalAngles = coneAnglesForSegment;
-                        break;
-
-                    case HPUIInteractorConeRayAngleSegment.LittleDistalSegment:
-                        estimatedConeRayAngles.LittleDistalAngles = coneAnglesForSegment;
-                        break;
-                    case HPUIInteractorConeRayAngleSegment.LittleIntermediateSegment:
-                        estimatedConeRayAngles.LittleIntermediateAngles = coneAnglesForSegment;
-                        break;
-                    case HPUIInteractorConeRayAngleSegment.LittleProximalSegment:
-                        estimatedConeRayAngles.LittleProximalAngles = coneAnglesForSegment;
-                        break;
-                }
-            }
-
-            try
-            {
-                callback.Invoke(estimatedConeRayAngles);
-            }
-            catch (Exception e)
-            {
-                Debug.LogError($"{e}");
-            }
-        }
-
-        /// <summary>
-        /// For a given segment, computes the List of <see cref="HPUIInteractorRayAngle">.
-        /// </summary>
-        /// <param name="segment">
-        ///   The <see cref="HPUIInteractorConeRayAngleSegment"/> for which cone
-        ///   angles are being computed.
-        /// </param>
-        /// <param name="interactionRecords">
-        ///   Each <see cref="InteractionDataRecord"/> in the list is expected to be the data collected during a single gesture event.
-        ///   The <see cref="InteractionDataRecord.segment">segment</see> of each record will be based on the interactableSegmentPairs
-        ///   passed in the constructor.
-        /// </param>
-        /// <param name="interactableSegmentPairs">
-        ///   The interactableSegmentPairs passed in the constructor.
-        /// </param>
-        protected virtual List<HPUIInteractorRayAngle> EstimateConeAnglesForSegment(HPUIInteractorConeRayAngleSegment segment, List<InteractionDataRecord> interactionRecords)
+        /// <inheritdoc />
+        protected override List<HPUIInteractorRayAngle> EstimateConeAnglesForSegment(HPUIInteractorConeRayAngleSegment segment)
         {
             Dictionary<(float, float), float> averageRayDistance = new();
             // For each interaction, get the frame with the shortest distance
@@ -242,24 +82,6 @@ namespace ubco.ovilab.HPUI.Components
             }
 
             return coneAnglesForSegment;
-        }
-
-        /// <summary>
-        /// Holds all the data collected for a single calibration event.
-        /// </summary>
-        protected struct InteractionDataRecord
-        {
-            // the frames captured, where each frame contains all the rays with data in that frame
-            // see <see cref="HPUIRayCastDetectionBaseLogic.RaycastDataRecord"/> for more info
-            public List<List<HPUIRayCastDetectionBaseLogic.RaycastDataRecord>> records;
-            // the segment for which the calibration is being performed
-            public HPUIInteractorConeRayAngleSegment segment;
-
-            public InteractionDataRecord(List<List<HPUIRayCastDetectionBaseLogic.RaycastDataRecord>> records, HPUIInteractorConeRayAngleSegment segment) : this()
-            {
-                this.records = records;
-                this.segment = segment;
-            }
         }
     }
 }
