+++ Runtime/Components/ConeRayAnglesEstimator/OnGestureConeRayEstimator.cs
@@ -1,10 +1,7 @@
 using System.Collections.Generic;
-using UnityEngine;
 using System;
 using UnityEngine.Assertions;
 using System.Linq;
-using System.Threading.Tasks;
-using System.Collections;
 using ubco.ovilab.HPUI.Interaction;
 
 namespace ubco.ovilab.HPUI.Components
@@ -58,14 +55,9 @@ namespace ubco.ovilab.HPUI.Components
     /// <see cref="HPUIFullRangeRayCastDetectionLogic.raycastData"/>
     /// </remarks>
     /// <seealso cref="EstimateConeRayAngles"/>
-    public class OnGestureConeRayEstimator
+    public class OnGestureConeRayEstimator: ConeRayEstimator 
     {
-        private HPUIInteractor interactor;
         private List<ConeRayAnglesEstimationPair> interactableSegmentPairs;
-        private HPUIFullRangeRayCastDetectionLogic fullRayDetectionLogic;
-        private HPUIInteractorFullRangeAngles fullRangeAngles;
-        private List<InteractionDataRecord> interactionRecords = new();
-        private List<List<HPUIRayCastDetectionBaseLogic.RaycastDataRecord>> currentInteractionData = new();
 
         /// <summary>
         /// Instantiates the estimator. This will subscribe to <see cref="HPUIInteractor.DetectionLogic"/>interactor.DetectionLogic</see>
@@ -83,40 +75,18 @@ namespace ubco.ovilab.HPUI.Components
         ///   If not true, interactableSegmentPairs should have atleast one entry for each
         ///   <see cref="HPUIInteractorConeRayAngleSegment"/>.
         /// </param>
-        public OnGestureConeRayEstimator(HPUIInteractor interactor, List<ConeRayAnglesEstimationPair> interactableSegmentPairs, bool ignoreMissingSegments=false)
+        public OnGestureConeRayEstimator(HPUIInteractor interactor, List<ConeRayAnglesEstimationPair> interactableSegmentPairs, bool ignoreMissingSegments=false): base(interactor)
         {
-            if (!(interactor.DetectionLogic is HPUIFullRangeRayCastDetectionLogic fullRayDetectionLogic))
-            {
-                throw new ArgumentException("Interactor is expected to have `HPUIFullRangeRayCastDetectionLogic` as the DetectionLogic.");
-            }
-
             if (!ignoreMissingSegments && interactableSegmentPairs.Select(el => el.segment).Distinct().Count() != Enum.GetValues(typeof(HPUIInteractorConeRayAngleSegment)).Length)
             {
                 throw new ArgumentException("Expecting all segments in interactableToSegmentMapping");
             }
 
-            this.interactor = interactor;
             this.interactableSegmentPairs = interactableSegmentPairs;
-            this.fullRayDetectionLogic = fullRayDetectionLogic;
-            this.fullRangeAngles = fullRayDetectionLogic.FullRangeRayAngles;
-
-            fullRayDetectionLogic.raycastData += RaycastDataCallback;
 
             foreach (IHPUIInteractable interactable in interactableSegmentPairs.Select(el => el.interactable).Distinct())
             {
                 interactable.GestureEvent.AddListener(OnGestureCallback);
-            }
-        }
-
-        /// <summary>
-        /// The callback used to get the data from the <see cref="HPUIFullRangeRayCastDetectionLogic.raycastData"/>.
-        /// </summary>
-        protected void RaycastDataCallback(List<HPUIRayCastDetectionBaseLogic.RaycastDataRecord> raycastDataRecords)
-        {
-            Assert.AreEqual(fullRangeAngles, ((HPUIFullRangeRayCastDetectionLogic)interactor.DetectionLogic).FullRangeRayAngles);
-            if (raycastDataRecords.Count > 0)
-            {
-                currentInteractionData.Add(raycastDataRecords);
             }
         }
 
@@ -138,132 +108,19 @@ namespace ubco.ovilab.HPUI.Components
             }
         }
 
-        /// <summary>
-        /// Stops the data collection and initates the estimation. Once done, callback will be invoked with the estimated asset.
-        /// </summary>
-        /// <remarks>
-        /// This will unsubscribe to <see cref="HPUIInteractor.DetectionLogic"/>interactor.DetectionLogic</see>
-        /// and the <see cref="IHPUIInteractable.GestureEvent">GestureEvent</see> of each
-        /// interactable it is tracking.
-        /// </remarks>
-        public void EstimateConeRayAngles(Action<HPUIInteractorConeRayAngles> callback)
+        /// <inheritdoc />
+        public override void EstimateConeRayAngles(Action<HPUIInteractorConeRayAngles> callback)
         {
-            HPUIInteractorConeRayAngles estimatedConeRayAngles = ScriptableObject.CreateInstance<HPUIInteractorConeRayAngles>();
-
-            fullRayDetectionLogic.raycastData -= RaycastDataCallback;
-
             foreach (IHPUIInteractable interactable in interactableSegmentPairs.Select(el => el.interactable).Distinct())
             {
                 interactable.GestureEvent.RemoveListener(OnGestureCallback);
             }
 
-            interactor.StartCoroutine(EstimationCoroutine(callback, estimatedConeRayAngles));
+            base.EstimateConeRayAngles(callback);
         }
 
-        /// <summary>
-        /// Coroutine that does the actual work of <see cref="EstimateConeRayAngles"/>.
-        /// </summary>
-        protected IEnumerator EstimationCoroutine(Action<HPUIInteractorConeRayAngles> callback, HPUIInteractorConeRayAngles estimatedConeRayAngles)
-        {
-            yield return null;
-            HPUIInteractorConeRayAngleSegment[] segments = (HPUIInteractorConeRayAngleSegment[])Enum.GetValues(typeof(HPUIInteractorConeRayAngleSegment));
-
-            Task<List<HPUIInteractorRayAngle>>[] tasks = new Task<List<HPUIInteractorRayAngle>>[segments.Length];
-
-            int i = 0;
-            foreach (HPUIInteractorConeRayAngleSegment segment in segments)
-            {
-                tasks[i++] = Task.Run(() => EstimateConeAnglesForSegment(segment, interactionRecords, interactableSegmentPairs));
-            }
-
-            while (tasks.Any(t => !t.IsCompleted))
-            {
-                IEnumerable<Exception> exceptions = tasks.Select(t => t.Exception).Where(t => t != null);
-                if(exceptions.Count() != 0)
-                {
-                    // Throwing the first thing that comes through.
-                    // MAYBE: Should all of them be processed somehow?
-                    throw exceptions.First();
-                }
-                yield return null;
-            }
-
-            for (i = 0; i < segments.Length; ++i)
-            {
-                HPUIInteractorConeRayAngleSegment segment = segments[i];
-                List<HPUIInteractorRayAngle> coneAnglesForSegment = tasks[i].Result;
-                switch (segment)
-                {
-                    case HPUIInteractorConeRayAngleSegment.IndexDistalSegment:
-                        estimatedConeRayAngles.IndexDistalAngles = coneAnglesForSegment;
-                        break;
-                    case HPUIInteractorConeRayAngleSegment.IndexIntermediateSegment:
-                        estimatedConeRayAngles.IndexIntermediateAngles = coneAnglesForSegment;
-                        break;
-                        case HPUIInteractorConeRayAngleSegment.IndexProximalSegment:
-                            estimatedConeRayAngles.IndexProximalAngles = coneAnglesForSegment;
-                            break;
-
-                        case HPUIInteractorConeRayAngleSegment.MiddleDistalSegment:
-                            estimatedConeRayAngles.MiddleDistalAngles = coneAnglesForSegment;
-                            break;
-                        case HPUIInteractorConeRayAngleSegment.MiddleIntermediateSegment:
-                            estimatedConeRayAngles.MiddleIntermediateAngles = coneAnglesForSegment;
-                            break;
-                        case HPUIInteractorConeRayAngleSegment.MiddleProximalSegment:
-                            estimatedConeRayAngles.MiddleProximalAngles = coneAnglesForSegment;
-                            break;
-
-                        case HPUIInteractorConeRayAngleSegment.RingDistalSegment:
-                            estimatedConeRayAngles.RingDistalAngles = coneAnglesForSegment;
-                            break;
-                        case HPUIInteractorConeRayAngleSegment.RingIntermediateSegment:
-                            estimatedConeRayAngles.RingIntermediateAngles = coneAnglesForSegment;
-                            break;
-                        case HPUIInteractorConeRayAngleSegment.RingProximalSegment:
-                            estimatedConeRayAngles.RingProximalAngles = coneAnglesForSegment;
-                            break;
-
-                        case HPUIInteractorConeRayAngleSegment.LittleDistalSegment:
-                            estimatedConeRayAngles.LittleDistalAngles = coneAnglesForSegment;
-                            break;
-                        case HPUIInteractorConeRayAngleSegment.LittleIntermediateSegment:
-                            estimatedConeRayAngles.LittleIntermediateAngles = coneAnglesForSegment;
-                            break;
-                        case HPUIInteractorConeRayAngleSegment.LittleProximalSegment:
-                            estimatedConeRayAngles.LittleProximalAngles = coneAnglesForSegment;
-                            break;
-                }
-            }
-
-            try
-            {
-                callback.Invoke(estimatedConeRayAngles);
-            }
-            catch(Exception e)
-            {
-                Debug.LogError($"{e}");
-            }
-        }
-
-        /// <summary>
-        /// For a given segment, computes the List of <see cref="HPUIInteractorRayAngle">.
-        /// </summary>
-        /// <param name="segment">
-        ///   The <see cref="HPUIInteractorConeRayAngleSegment"/> for which cone
-        ///   angles are being computed.
-        /// </param>
-        /// <param name="interactionRecords">
-        ///   Each <see cref="InteractionDataRecord"/> in the list is expected to be the data collected during a single gesture event.
-        ///   The <see cref="InteractionDataRecord.segment">segment</see> of each record will be based on the interactableSegmentPairs
-        ///   passed in the constructor.
-        /// </param>
-        /// <param name="interactableSegmentPairs">
-        ///   The interactableSegmentPairs passed in the constructor.
-        /// </param>
-        protected virtual List<HPUIInteractorRayAngle> EstimateConeAnglesForSegment(HPUIInteractorConeRayAngleSegment segment,
-                                                                                    List<InteractionDataRecord> interactionRecords,
-                                                                                    List<ConeRayAnglesEstimationPair> interactableSegmentPairs)
+        /// <inheritdoc />
+        protected override List<HPUIInteractorRayAngle> EstimateConeAnglesForSegment(HPUIInteractorConeRayAngleSegment segment)
         {
             List<HPUIRayCastDetectionBaseLogic.RaycastDataRecord> filteredInteractionRecords = new();
 
@@ -315,21 +172,6 @@ namespace ubco.ovilab.HPUI.Components
                 .ToList();
 
             return coneAnglesForSegment;
-        }
-
-        /// <summary>
-        /// Holds all the data collected for a single gesture event.
-        /// </summary>
-        protected struct InteractionDataRecord
-        {
-            public List<List<HPUIRayCastDetectionBaseLogic.RaycastDataRecord>> records;
-            public HPUIInteractorConeRayAngleSegment segment;
-
-            public InteractionDataRecord(List<List<HPUIRayCastDetectionBaseLogic.RaycastDataRecord>> records, HPUIInteractorConeRayAngleSegment segment) : this()
-            {
-                this.records = records;
-                this.segment = segment;
-            }
         }
     }
 }
